pyproject.toml
[project]
name = "webcam-spatial-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "mediapipe",
    "opencv-python",
    "python-dotenv",
]
